set up core functionality for EditTask and onCheckBoxClicked in MainActivity

+++ app/src/main/java/com/example/taskmanager/Task.java
@@ -1,13 +1,13 @@
 package com.example.taskmanager;
 
 public class Task {
-    private long id;
+    private int id; // I wanted to keep it long but view.setId() only accepts int values
     private String description;
     private String created_at;
     private String complete_time;
     private boolean is_completed;
 
-    public Task(long id, String description, String created_at, String complete_time, boolean is_completed) {
+    public Task(int id, String description, String created_at, String complete_time, boolean is_completed) {
         this.id = id;
         this.description = description;
         this.created_at = created_at;
@@ -15,7 +15,7 @@ public class Task {
         this.is_completed = is_completed;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
